Remove parent <p>  from <Link /> in Nav.js
Hello I`m submitting this PR to bring to light a small bug that none the less cause me a lot of confusion in this tutorial "https://snipcart.com/blog/react-ecommerce-tutorial" . I`m very new to React, Nextjs and this is my first time using Snipcart so bear in my that this feedback is coming from a humble n00b who may have missed some important detail on his end. I wanted to bring it up because it may detract people from using Snipcart as it almost did for me.

1) When I initially implemented Step 4 the cart appeared with no issues. Until I played around with it and removed some items, added others. Eventually I received these errors:
-  "Error: Hydration failed because the initial UI does not match what was rendered on the server. Warning: Did not expect server HTML to contain a <div> in <div>."
-  "Error: There was an error while hydrating. Because the error happened outside of a Suspense boundary, the entire root will switch to client rendering."

When did the command `npm run dev` fixed the issue. This might be useful to put in as well

I was able to address the issue by taking <Link/> out of <p> in Nav.js

Before

...
Before
...
<p className={styles.title}>
  <Link href="/">
   <a>Snipcart Store</a>
  </Link>
 </p>
...

After:
...

  <Link href="/">
   <a>Snipcart Store</a>
  </Link>

...

Based on my reading this error used to only come up in the console but now completely stops the app from running after React 18 launched which was the same month that this article came out. I`d suggest that you update this article to take that into account.

Another point I'd like to add is that the article was helpful enough to include a l"Follow along with the commit in GitHub!" link on step 4 but it directed to this commit from 2020 https://github.com/snipcart/next-snipcart-store/commit/556e6f4af1d1ec62b49e064b70e0f1a9fb198c25 which had a directory that looked nothing like what I had up to that point following the tutorial. 

Just wanted to bring this up and suggest an updated to this blog  to represent the changes from React 18 as it was my first introduction to Snipcart.
Anyways I look forward to implementing this technology to my next project! 

Cheers
Phil
https://philjgray.ca/